refactor: use the app config
